Extract section heading symbol partial
