+++ src/CodeIndex/Indexer/Symbols/SymbolExtractor.cs
@@ -22,12 +22,6 @@ public static partial class SymbolExtractor
     public const int FunctionalLanguageContractVersion = 2;
     public const int DynamicLanguageContractVersion = 2;
     public const int SystemsLanguageContractVersion = 2;
-    private static readonly Regex CSharpRegionRegex = new(
-        @"^\s*#region(?:\s+(?<name>.*\S))?\s*$",
-        RegexOptions.Compiled | RegexOptions.CultureInvariant);
-    private static readonly Regex JavaScriptTypeScriptModuleDocRegex = new(
-        @"@module(?:\s+(?<name>[^\s*]+))?",
-        RegexOptions.Compiled | RegexOptions.CultureInvariant);
     private static readonly Regex ShellHeredocRedirectRegex = new(
         @"(?<!<)<<-?(?!<)\s*(?:""(?<dq>[^""]+)""|'(?<sq>[^']+)'|\\?(?<bare>[A-Za-z_][A-Za-z0-9_]*))",
         RegexOptions.Compiled | RegexOptions.CultureInvariant);
@@ -4731,83 +4725,6 @@ public static partial class SymbolExtractor
             return index + 1;
 
         return ~index;
-    }
-
-    private static void ExtractSectionHeadingSymbols(long fileId, string lang, string[] lines, List<SymbolRecord> symbols)
-    {
-        if (lang == "csharp")
-        {
-            for (var i = 0; i < lines.Length; i++)
-            {
-                var match = CSharpRegionRegex.Match(lines[i]);
-                if (!match.Success)
-                    continue;
-
-                AddHeadingSymbol(fileId, lines, symbols, i, match.Groups["name"].Value.Trim(), "#region");
-            }
-        }
-        else if (lang == "python")
-        {
-            TryAddPythonModuleDocstringHeading(fileId, lines, symbols);
-        }
-        else
-        {
-            for (var i = 0; i < lines.Length; i++)
-            {
-                var match = JavaScriptTypeScriptModuleDocRegex.Match(lines[i]);
-                if (!match.Success)
-                    continue;
-
-                AddHeadingSymbol(fileId, lines, symbols, i, match.Groups["name"].Value.Trim(), "@module");
-            }
-        }
-    }
-
-    private static void TryAddPythonModuleDocstringHeading(long fileId, string[] lines, List<SymbolRecord> symbols)
-    {
-        for (var i = 0; i < lines.Length; i++)
-        {
-            var trimmed = lines[i].TrimStart();
-            if (trimmed.Length == 0 || trimmed.StartsWith("#", StringComparison.Ordinal))
-                continue;
-
-            var quote = trimmed.StartsWith("\"\"\"", StringComparison.Ordinal) ? "\"\"\"" :
-                trimmed.StartsWith("'''", StringComparison.Ordinal) ? "'''" : null;
-            if (quote == null)
-                return;
-
-            var name = trimmed[quote.Length..].Trim();
-            if (name.EndsWith(quote, StringComparison.Ordinal))
-                name = name[..^quote.Length].Trim();
-            AddHeadingSymbol(fileId, lines, symbols, i, name, "module docstring");
-            return;
-        }
-    }
-
-    private static void AddHeadingSymbol(
-        long fileId,
-        string[] lines,
-        List<SymbolRecord> symbols,
-        int lineIndex,
-        string name,
-        string fallbackName)
-    {
-        var lineNumber = lineIndex + 1;
-        AddSymbolRecord(
-            symbols,
-            null,
-            lineNumber,
-            new SymbolRecord
-            {
-                FileId = fileId,
-                Kind = "heading",
-                Name = string.IsNullOrWhiteSpace(name) ? fallbackName : name,
-                Line = lineNumber,
-                StartLine = lineNumber,
-                EndLine = lineNumber,
-                Signature = lines[lineIndex].Trim(),
-            },
-            lines[lineIndex]);
     }
 
     private static void ExtractSqlGeneratedColumnSymbols(long fileId, string[] lines, string[] structuralLines, List<SymbolRecord> symbols)
